Refactor widget postMessage API.
