+++ src/WidgetMessaging.js
@@ -19,38 +19,33 @@ limitations under the License.
 * spec. details / documentation.
 */
 
-import URL from 'url';
-import dis from './dispatcher';
-import MatrixPostMessageApi from './MatrixPostMessageApi';
-import IntegrationManager from './IntegrationManager';
+import FromWidgetPostMessageApi from './FromWidgetPostMessageApi';
+import ToWidgetPostMessageApi from './ToWidgetPostMessageApi';
 
-const WIDGET_API_VERSION = '0.0.1'; // Current API version
-const SUPPORTED_WIDGET_API_VERSIONS = [
-    '0.0.1',
-];
-const INBOUND_API_NAME = 'fromWidget';
+if (!global.mxFromWidgetMessaging) {
+    global.mxFromWidgetMessaging = new FromWidgetPostMessageApi();
+    global.mxFromWidgetMessaging.start();
+}
+if (!global.mxToWidgetMessaging) {
+    global.mxToWidgetMessaging = new ToWidgetPostMessageApi();
+    global.mxToWidgetMessaging.start();
+}
+
 const OUTBOUND_API_NAME = 'toWidget';
 
-if (!global.mxWidgetMessagingListenerCount) {
-    global.mxWidgetMessagingListenerCount = 0;
-}
-if (!global.mxWidgetMessagingMessageEndpoints) {
-    global.mxWidgetMessagingMessageEndpoints = [];
-}
-
-export default class WidgetMessaging extends MatrixPostMessageApi {
-    constructor(widgetId, targetWindow) {
-        super(targetWindow);
+export default class WidgetMessaging {
+    constructor(widgetId, widgetUrl, target) {
         this.widgetId = widgetId;
-
-        this.startListening = this.startListening.bind(this);
-        this.stopListening = this.stopListening.bind(this);
-        this.onMessage = this.onMessage.bind(this);
+        this.widgetUrl = widgetUrl;
+        this.target = target;
+        this.fromWidget = global.mxFromWidgetMessaging;
+        this.toWidget = global.mxToWidgetMessaging;
+        this.start();
     }
 
-    exec(action) {
-        return super.exec(action).then((data) => {
-            // check for errors and reject if found
+    messageToWidget(action) {
+        return this.toWidget.exec(action, this.target).then((data) => {
+            // Check for errors and reject if found
             if (data.response === undefined) { // null is valid
                 throw new Error("Missing 'response' field");
             }
@@ -65,196 +60,9 @@ export default class WidgetMessaging extends MatrixPostMessageApi {
                 // We can't aggressively sanitize [A-z0-9] since it might be a translation.
                 throw new Error(msg);
             }
-            // return the response field for the request
+            // Return the response field for the request
             return data.response;
         });
-    }
-
-    /**
-     * Register widget message event listeners
-     */
-    startListening() {
-        if (global.mxWidgetMessagingListenerCount === 0) {
-            // Start postMessage API listener
-            this.start();
-            // Start widget specific listener
-            window.addEventListener("message", this.onMessage, false);
-        }
-        global.mxWidgetMessagingListenerCount += 1;
-    }
-
-    /**
-     * De-register widget message event listeners
-     */
-    stopListening() {
-        global.mxWidgetMessagingListenerCount -= 1;
-        if (global.mxWidgetMessagingListenerCount === 0) {
-            // Stop widget specific listener
-            window.removeEventListener("message", this.onMessage, false);
-            // Stop postMessage API listener
-            this.stop();
-        }
-        if (global.mxWidgetMessagingListenerCount < 0) {
-            // Make an error so we get a stack trace
-            const e = new Error(
-                "WidgetMessaging: mismatched startListening / stopListening detected." +
-                " Negative count",
-            );
-            console.error(e);
-        }
-    }
-
-    /**
-     * Register a widget endpoint for trusted postMessage communication
-     * @param {string} widgetId    Unique widget identifier
-     * @param {string} endpointUrl Widget wurl origin (protocol + (optional port) + host)
-     */
-    addEndpoint(widgetId, endpointUrl) {
-        const u = URL.parse(endpointUrl);
-        if (!u || !u.protocol || !u.host) {
-            console.warn("Invalid origin:", endpointUrl);
-            return;
-        }
-
-        const origin = u.protocol + '//' + u.host;
-        const endpoint = new WidgetMessageEndpoint(widgetId, origin);
-        if (global.mxWidgetMessagingMessageEndpoints) {
-            if (global.mxWidgetMessagingMessageEndpoints.some(function(ep) {
-                return (ep.widgetId === widgetId && ep.endpointUrl === endpointUrl);
-            })) {
-                // Message endpoint already registered
-                console.warn("Endpoint already registered");
-                return;
-            } else {
-                // console.warn(`Adding widget messaging endpoint for ${widgetId}`);
-                global.mxWidgetMessagingMessageEndpoints.push(endpoint);
-            }
-        }
-    }
-
-    /**
-     * De-register a widget endpoint from trusted communication sources
-     * @param  {string} widgetId Unique widget identifier
-     * @param  {string} endpointUrl Widget wurl origin (protocol + (optional port) + host)
-     * @return {boolean} True if endpoint was successfully removed
-     */
-    removeEndpoint(widgetId, endpointUrl) {
-        const u = URL.parse(endpointUrl);
-        if (!u || !u.protocol || !u.host) {
-            console.warn("Invalid origin");
-            return;
-        }
-
-        const origin = u.protocol + '//' + u.host;
-        if (global.mxWidgetMessagingMessageEndpoints && global.mxWidgetMessagingMessageEndpoints.length > 0) {
-            const length = global.mxWidgetMessagingMessageEndpoints.length;
-            global.mxWidgetMessagingMessageEndpoints = global.mxWidgetMessagingMessageEndpoints.
-                filter(function(endpoint) {
-                return (endpoint.widgetId != widgetId || endpoint.endpointUrl != origin);
-            });
-            return (length > global.mxWidgetMessagingMessageEndpoints.length);
-        }
-        return false;
-    }
-
-    /**
-     * Handle widget postMessage events
-     * @param  {Event} event Event to handle
-     * @return {undefined}
-     */
-    onMessage(event) {
-        if (!event.origin) { // Handle chrome
-            event.origin = event.originalEvent.origin;
-        }
-
-        // Event origin is empty string if undefined
-        if (
-            event.origin.length === 0 ||
-            !this.trustedEndpoint(event.origin) ||
-            event.data.api !== INBOUND_API_NAME ||
-            !event.data.widgetId
-        ) {
-            return; // don't log this - debugging APIs like to spam postMessage which floods the log otherwise
-        }
-
-        const action = event.data.action;
-        const widgetId = event.data.widgetId;
-        if (action === 'content_loaded') {
-            dis.dispatch({
-                action: 'widget_content_loaded',
-                widgetId: widgetId,
-            });
-            this.sendResponse(event, {success: true});
-        } else if (action === 'supported_api_versions') {
-            this.sendResponse(event, {
-                api: INBOUND_API_NAME,
-                supported_versions: SUPPORTED_WIDGET_API_VERSIONS,
-            });
-        } else if (action === 'api_version') {
-            this.sendResponse(event, {
-                api: INBOUND_API_NAME,
-                version: WIDGET_API_VERSION,
-            });
-        } else if (action === 'm.sticker') {
-            dis.dispatch({action: 'm.sticker', data: event.data.widgetData, widgetId: event.data.widgetId});
-        } else if (action === 'integration_manager_open') {
-            // Close the stickerpicker
-            dis.dispatch({action: 'stickerpicker_close'});
-            // Open the integration manager
-            const data = event.data.widgetData;
-            const integType = (data && data.integType) ? data.integType : null;
-            const integId = (data && data.integId) ? data.integId : null;
-            IntegrationManager.open(integType, integId);
-        } else {
-            console.warn("Widget postMessage event unhandled");
-            this.sendError(event, {message: "The postMessage was unhandled"});
-        }
-    }
-
-    /**
-     * Check if message origin is registered as trusted
-     * @param  {string} origin PostMessage origin to check
-     * @return {boolean}       True if trusted
-     */
-    trustedEndpoint(origin) {
-        if (!origin) {
-            return false;
-        }
-
-        return global.mxWidgetMessagingMessageEndpoints.some((endpoint) => {
-            return endpoint.endpointUrl === origin;
-        });
-    }
-
-    /**
-     * Send a postmessage response to a postMessage request
-     * @param  {Event} event  The original postMessage request event
-     * @param  {Object} res   Response data
-     */
-    sendResponse(event, res) {
-        const data = JSON.parse(JSON.stringify(event.data));
-        data.response = res;
-        event.source.postMessage(data, event.origin);
-    }
-
-    /**
-     * Send an error response to a postMessage request
-     * @param  {Event} event        The original postMessage request event
-     * @param  {string} msg         Error message
-     * @param  {Error} nestedError  Nested error event (optional)
-     */
-    sendError(event, msg, nestedError) {
-        console.error("Action:" + event.data.action + " failed with message: " + msg);
-        const data = JSON.parse(JSON.stringify(event.data));
-        data.response = {
-            error: {
-                message: msg,
-            },
-        };
-        if (nestedError) {
-            data.response.error._error = nestedError;
-        }
-        event.source.postMessage(data, event.origin);
     }
 
     /**
@@ -262,11 +70,13 @@ export default class WidgetMessaging extends MatrixPostMessageApi {
      * @return {Promise} To be resolved with screenshot data when it has been generated
      */
     getScreenshot() {
-        return this.exec({
+        console.warn('Requesting screenshot for', this.widgetId);
+        return this.messageToWidget({
                 api: OUTBOUND_API_NAME,
                 action: "screenshot",
-            }).then((response) => response.screenshot)
-            .catch((error) => new Error("Failed to get screenshot: " + error.message));
+            })
+            .catch((error) => new Error("Failed to get screenshot: " + error.message))
+            .then((response) => response.screenshot);
     }
 
     /**
@@ -274,30 +84,22 @@ export default class WidgetMessaging extends MatrixPostMessageApi {
      * @return {Promise} To be resolved with an array of requested widget capabilities
      */
     getCapabilities() {
-        return this.exec({
+        console.warn('Requesting capabilities for', this.widgetId);
+        return this.messageToWidget({
                 api: OUTBOUND_API_NAME,
                 action: "capabilities",
-            }).then((response) => response.capabilities);
+            }).then((response) => {
+                console.warn('Got capabilities for', this.widgetId, response.capabilities);
+                return response.capabilities;
+            });
     }
-}
 
-/**
- * Represents mapping of widget instance to URLs for trusted postMessage communication.
- */
-class WidgetMessageEndpoint {
-    /**
-     * Mapping of widget instance to URL for trusted postMessage communication.
-     * @param  {string} widgetId    Unique widget identifier
-     * @param  {string} endpointUrl Widget wurl origin.
-     */
-    constructor(widgetId, endpointUrl) {
-        if (!widgetId) {
-            throw new Error("No widgetId specified in widgetMessageEndpoint constructor");
-        }
-        if (!endpointUrl) {
-            throw new Error("No endpoint specified in widgetMessageEndpoint constructor");
-        }
-        this.widgetId = widgetId;
-        this.endpointUrl = endpointUrl;
+
+    start() {
+        this.fromWidget.addEndpoint(this.widgetId, this.widgetUrl);
+    }
+
+    stop() {
+        this.fromWidget.removeEndpoint(this.widgetId, this.widgetUrl);
     }
 }
